requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
faiss-cpu
python-dotenv
google-generativeai
